Move function 'parse_formatted_text' to utils.rs
It is planned to use this function in other places like for parsing
the ToS text. Consequently, its companion function
'convert_to_markup' is also moved to utils.rs.
